+++ alacritty/src/config/mouse.rs
@@ -8,6 +8,7 @@ use crate::config::ui_config;
 #[derive(ConfigDeserialize, Serialize, Default, Clone, Debug, PartialEq, Eq)]
 pub struct Mouse {
     pub hide_when_typing: bool,
+    pub zoom_using_wheel: bool,
     #[serde(skip_serializing)]
     pub bindings: MouseBindings,
 }