+++ alacritty/src/event.rs
@@ -1974,7 +1974,13 @@ impl input::Processor<EventProxy, ActionContext<'_, Notifier, EventProxy>> {
                     },
                     WindowEvent::MouseWheel { delta, phase, .. } => {
                         self.ctx.window().set_mouse_visible(true);
-                        self.mouse_wheel_input(delta, phase);
+                        if self.ctx.config.mouse.zoom_using_wheel
+                            && self.ctx.modifiers().state().control_key()
+                        {
+                            self.mouse_wheel_zoom(delta, phase);
+                        } else {
+                            self.mouse_wheel_input(delta, phase);
+                        }
                     },
                     WindowEvent::Touch(touch) => self.touch(touch),
                     WindowEvent::Focused(is_focused) => {